array: avoid custom Thetis branch

+++ turbine_adapt/parse.py
@@ -12,7 +12,7 @@ class Parser(object):
             'level': {'type': int, 'msg': """
                 Resolution level of initial mesh (default {:d}).
                 """},
-            'end_time': {'type': int, 'msg': """
+            'end_time': {'type': float, 'msg': """
                 Simulation end time in seconds (default {:.1f}).
                 """},
             'num_meshes': {'type': int, 'msg': """
